[jit] move script-related implementation out of torch/jit/__init__.py
See the bottom of this stack for context.

[ghstack-poisoned]
