+++ torch/jit/_state.py
@@ -7,6 +7,7 @@ functionalities in `torch.jit`.
 """
 import torch
 import os
+import weakref
 
 
 class EnabledProxy:
@@ -55,3 +56,55 @@ def enable():
 # live in here. It's defined in Python because doing in cpp creates static
 # destruction order issues.
 _python_cu = torch._C.CompilationUnit()
+
+
+# qualified_name => ScriptClass mapping
+_script_classes = {}
+
+def _add_script_class(cls, name):
+    cls.__torch_script_class__ = True
+    global _script_classes
+    _script_classes[name] = cls
+
+
+def _get_script_class(name):
+    global _script_classes
+    if name not in _script_classes:
+        return None
+    return _script_classes[name]
+
+
+# Caching: we currently cache compilation of free functions and overloaded functions.
+# To cache free functions we hold a weak ref to the function object and
+# map to the compiled fn's qualified name.
+# To cache overloaded functions we hold a weak ref to the function obj and
+# map to all of its overloaded compiled fns.
+# In the future we could consider caching more types of objects so that
+# aliasing is preserved across separate compilations of the same object.
+
+_jit_caching_layer = weakref.WeakKeyDictionary()
+_jit_function_overload_caching = weakref.WeakKeyDictionary()
+
+def _try_get_jit_cached_overloads(key):
+    qual_names = _jit_function_overload_caching.get(key, None)
+    if qual_names:
+        return [_python_cu.find_function(qual_name) for qual_name in qual_names]
+    else:
+        return None
+
+def _set_jit_overload_cache(key, compiled_fns):
+    _jit_function_overload_caching[key] = [fn.qualified_name for fn in compiled_fns]
+
+def _try_get_jit_cached_function(key):
+    if getattr(key, "__disable_jit_function_caching__", False) is True:
+        return None
+    qual_name = _jit_caching_layer.get(key, None)
+    if qual_name:
+        return _python_cu.find_function(qual_name)
+    else:
+        return None
+
+def _set_jit_function_cache(key, value):
+    # only free functions currently supported
+    assert isinstance(value, torch.jit.ScriptFunction)
+    _jit_caching_layer[key] = value.qualified_name
